+++ arancino/arancino_conf.py
@@ -34,9 +34,15 @@ redis_dts = {'host': 'localhost',
 
 #devicestore
 redis_dvs = {'host': 'localhost',
-         'port': 6380,
+         'port': 6379,
          'dcd_resp': True,  #decode response
-         'db': 0}
+         'db': 2}
+
+#datastore reserved keys
+redis_dts_rsvd = {'host': 'localhost',
+         'port': 6379,
+         'dcd_resp': True,  #decode response
+         'db': 2}
 
 
 #cycle interval time
@@ -81,6 +87,6 @@ def __get_file_handler():
 
 logger = logging.getLogger(__name)
 
-logger.setLevel(logging.INFO)
-#logger.addHandler(__get_console_handler())
+logger.setLevel(logging.DEBUG)
+logger.addHandler(__get_console_handler())
 logger.addHandler(__get_file_handler())