+++ arancino/arancino_datastore.py
@@ -25,9 +25,14 @@ class ArancinoDataStore():
 
     def __init__(self):
 
+        #data store
         self.__redis_pool_dts = redis.ConnectionPool(host=conf.redis_dts['host'], port=conf.redis_dts['port'], db=conf.redis_dts['db'], decode_responses=conf.redis_dts['dcd_resp'])
+
+        #device store
         self.__redis_pool_dvs = redis.ConnectionPool(host=conf.redis_dvs['host'], port=conf.redis_dvs['port'], db=conf.redis_dvs['db'], decode_responses=conf.redis_dvs['dcd_resp'])
 
+        #data store (reserved keys)
+        self.__redis_pool_dts_rsvd = redis.ConnectionPool(host=conf.redis_dts_rsvd['host'], port=conf.redis_dts_rsvd['port'], db=conf.redis_dts_rsvd['db'], decode_responses=conf.redis_dts_rsvd['dcd_resp'])
 
     def getDataStore(self):
         """
@@ -36,13 +41,6 @@ class ArancinoDataStore():
         :return:
         """
 
-        #datastore configuration
-        #dts_host = conf.redis_dts['host']
-        #dts_port = conf.redis_dts['port']
-        #dts_db = conf.redis_dts['db']
-        #dts_dcd_resp = conf.redis_dts['dcd_resp']
-
-        #__redis_pool_dts = redis.ConnectionPool(host=dts_host, port=dts_port, db=dts_db, decode_responses=dts_dcd_resp)
         return redis.Redis(connection_pool=self.__redis_pool_dts)
 
 
@@ -52,12 +50,13 @@ class ArancinoDataStore():
             manage configurations of Arancino Devices.
         :return:
         """
-
-        #devicestore configuration
-        #dvs_host = conf.redis_dvs['host']
-        #dvs_port = conf.redis_dvs['port']
-        #dvs_db = conf.redis_dvs['db']
-        #dvs_dcd_resp = conf.redis_dvs['dcd_resp']
-
-        #__redis_pool_dvs = redis.ConnectionPool(host=dvs_host, port=dvs_port, db=dvs_db, decode_responses=dvs_dcd_resp)
         return redis.Redis(connection_pool=self.__redis_pool_dvs)
+
+    def getDataStoreRsvd(self):
+        """
+        Gets a redis client from a connection pool. This client is used to
+            manage reserved keys, and/or persistent application keys.
+        :return:
+        """
+
+        return redis.Redis(connection_pool=self.__redis_pool_dts_rsvd)