+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 #### v 0.1.5 - YYYY.MM.DD
 
+* Store application values persistently
+
 #### v 0.1.4 - 2019.04.17
 * Setup.py now includes extras/*
 * Log info and error enhancement. #25